(added) command test using Mock
