--- tests/__init__.py
@@ -1,6 +1,0 @@
-"""
-Import test files here for easier execution
-"""
-
-from tests.db_tests import TestDatabase
-from tests.command_tests import TestCommand
