--- tests/command_tests.py
@@ -1,27 +1,0 @@
-import unittest
-
-class TestCommand(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-    def test_start(self):
-        pass
-
-    def test_help(self):
-        pass
-
-    def test_set(self):
-        pass
-
-    def test_getRepos(self):
-        pass
-
-    def test_setRepo(self):
-        pass
-
-    def test_issues(self):
-        pass
